Move back into dependencies.cpp
